[flang][NFC] Move new code to right place
Some new code was added to flang/Semantics that only depends on
facilities in flang/Evaluate.  Move it into Evaluate and clean up
some minor stylistic problems.
